### 2023-01-06 #5050 @editorialbot
[PRE REVIEW]: unfold: removing the barriers to sharing and reproducing prospective life-cycle assessment databases
**Submitting author:** @romainsacchi (<a href="http://orcid.org/0000-0003-1440-0905">Romain Sacchi</a>)
**Repository:** https://github.com/polca/unfold
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
