### 2023-01-06 #5050 @kyleniemeyer
Hi @romainsacchi, due to the nature of your submission, the editorial board is going to review it a bit further to see if it meets our requirements for [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) and [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). We should hopefully get back to you within about a week or so. Thanks!
