+++ paper.bib
@@ -14,6 +14,7 @@ year = {2022}
 
 @article{Wernet:2016,
 author = {Wernet, G. and Bauer, C. and Steubing, B. and Reinhard, J. and Moreno-Ruiz, E. and Weidema, B.},
+doi = {10.1007/s11367-016-1087-8},
 journal = {The International Journal of Life Cycle Assessment},
 number = {9},
 pages = {1218--1230},
@@ -25,7 +26,7 @@ year = {2016}
 
 @article{Mendoza:2018,
 author = {{Mendoza Beltran}, Angelica and Cox, Brian and Mutel, Chris and van Vuuren, Detlef and Vivanco, David Font and Deetman, Sebastiaan and Edelenbosch, Oreane and Guin{\'{e}}e, Jeroen and Tukker, Arnold},
-doi = {https://doi.org/10.1111/jiec.12825},
+doi = {10.1111/jiec.12825},
 journal = {Journal of Industrial Ecology},
 mendeley-groups = {Carculator},
 title = {{When the Background Matters: Using Scenarios from Integrated Assessment Models in Prospective Life Cycle Assessment}},
@@ -61,7 +62,7 @@ year = {2021}
 @article{Steubing:2020,
 abstract = {The Activity Browser is an open source software for advanced Life Cycle Assessment (LCA). The Activity Browser provides a graphical user interface (GUI) to the brightway LCA framework and makes common tasks such as managing projects and databases, modeling life cycle inventories, and analyzing LCA results easier and more intuitive. In addition, it provides advanced features for LCA modeling and data analyses and thus facilitates state-of-the-art LCA research. It can be extended to implement novel LCA modeling approaches and analyses as needed.},
 author = {Steubing, Bernhard and de Koning, Daniel and Haas, Adrian and Mutel, Christopher Lucien},
-doi = {https://doi.org/10.1016/j.simpa.2019.100012},
+doi = {10.1016/j.simpa.2019.100012},
 issn = {2665-9638},
 journal = {Software Impacts},
 keywords = { Brightway, Scenario modeling,Life Cycle Assessment (LCA)},
