+++ unfold/data_cleaning.py
@@ -200,7 +200,7 @@ def check_for_duplicates(db: List[dict], data: List[dict]) -> List[dict]:
         (x["name"].lower(), x["reference product"].lower(), x["location"])
         for x in data
         if (x["name"].lower(), x["reference product"].lower(), x["location"])
-        in db_names
+           in db_names
     ]
 
     if len(already_exist) > 0:
@@ -249,7 +249,7 @@ def add_product_field_to_exchanges(data: List[dict], db: List[dict]) -> List[dic
         for exchange in dataset["exchanges"]:
             if exchange["type"] == "technosphere":
                 # Check if the field 'product' is present
-                if not "product" in exchange:
+                if "product" not in exchange:
                     exchange["product"] = correct_product_field(exchange, data, db)
 
     # Add a `code` field if missing
@@ -329,9 +329,9 @@ def correct_fields_format(data: list, name: str) -> list:
                 dataset["parameters"] = [dataset["parameters"]]
 
             if (
-                dataset["parameters"] is None
-                or dataset["parameters"] == {}
-                or dataset["parameters"] == []
+                    dataset["parameters"] is None
+                    or dataset["parameters"] == {}
+                    or dataset["parameters"] == []
             ):
                 del dataset["parameters"]
 
@@ -352,5 +352,54 @@ def correct_fields_format(data: list, name: str) -> list:
 
         if not dataset.get("database"):
             dataset["database"] = name
+
+    return data
+
+
+def check_mandatory_fields(data: list) -> list:
+    """
+    Check that the mandatory fields are present.
+    :param data: list of dictionaries representing the inventories to be imported
+    :raises ValueError: if a mandatory field is missing
+    :return: list of dictionaries representing the inventories to be imported
+    """
+
+    dataset_fields = [
+        "name",
+        "reference product",
+        "location",
+        "unit",
+        "exchanges",
+    ]
+
+    missing_fields = []
+
+    for dataset in data:
+        for field in dataset_fields:
+            if field not in dataset:
+                if field in ["reference product", "location", "unit", "name"] and "exchanges" in dataset:
+                    for exc in dataset["exchanges"]:
+                        if exc["type"] == "production":
+                            if field == "reference product":
+                                dataset[field] = exc["product"]
+                            else:
+                                dataset[field] = exc[field]
+                else:
+                    missing_fields.append([
+                        dataset.get("name", "unknown"),
+                        dataset.get("reference product", "unknown"),
+                        dataset.get("location", "unknown"),
+                        field
+                    ])
+
+    if missing_fields:
+        # print in prettytable the list of missing fields
+        table = PrettyTable()
+        table.field_names = ["Dataset", "Reference product", "Location", "Missing field"]
+        for row in missing_fields[:10]:
+            table.add_row(row)
+        print(table)
+        raise ValueError("Some mandatory fields are missing in the database."
+                         "Ten first missing fields are displayed above.")
 
     return data
