+++ unfold/data_cleaning.py
@@ -200,7 +200,7 @@ def check_for_duplicates(db: List[dict], data: List[dict]) -> List[dict]:
         (x["name"].lower(), x["reference product"].lower(), x["location"])
         for x in data
         if (x["name"].lower(), x["reference product"].lower(), x["location"])
-           in db_names
+        in db_names
     ]
 
     if len(already_exist) > 0:
@@ -329,9 +329,9 @@ def correct_fields_format(data: list, name: str) -> list:
                 dataset["parameters"] = [dataset["parameters"]]
 
             if (
-                    dataset["parameters"] is None
-                    or dataset["parameters"] == {}
-                    or dataset["parameters"] == []
+                dataset["parameters"] is None
+                or dataset["parameters"] == {}
+                or dataset["parameters"] == []
             ):
                 del dataset["parameters"]
 
@@ -377,7 +377,10 @@ def check_mandatory_fields(data: list) -> list:
     for dataset in data:
         for field in dataset_fields:
             if field not in dataset:
-                if field in ["reference product", "location", "unit", "name"] and "exchanges" in dataset:
+                if (
+                    field in ["reference product", "location", "unit", "name"]
+                    and "exchanges" in dataset
+                ):
                     for exc in dataset["exchanges"]:
                         if exc["type"] == "production":
                             if field == "reference product":
@@ -385,21 +388,30 @@ def check_mandatory_fields(data: list) -> list:
                             else:
                                 dataset[field] = exc[field]
                 else:
-                    missing_fields.append([
-                        dataset.get("name", "unknown"),
-                        dataset.get("reference product", "unknown"),
-                        dataset.get("location", "unknown"),
-                        field
-                    ])
+                    missing_fields.append(
+                        [
+                            dataset.get("name", "unknown"),
+                            dataset.get("reference product", "unknown"),
+                            dataset.get("location", "unknown"),
+                            field,
+                        ]
+                    )
 
     if missing_fields:
         # print in prettytable the list of missing fields
         table = PrettyTable()
-        table.field_names = ["Dataset", "Reference product", "Location", "Missing field"]
+        table.field_names = [
+            "Dataset",
+            "Reference product",
+            "Location",
+            "Missing field",
+        ]
         for row in missing_fields[:10]:
             table.add_row(row)
         print(table)
-        raise ValueError("Some mandatory fields are missing in the database."
-                         "Ten first missing fields are displayed above.")
+        raise ValueError(
+            "Some mandatory fields are missing in the database."
+            "Ten first missing fields are displayed above."
+        )
 
     return data
