### 2023-01-26 #5050 @romainsacchi
@fraukewiese Sorry about my late answer. Regarding the missing DOIs, [I have added them in the meanwhile](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05050/joss.05050/10.21105.joss.05050.pdf). Only two references miss them now: a Github repo and a web page. It is not possible to get a DOI for these resources as far as I know.
