### 2023-01-26 #5050 @romainsacchi
@fraukewiese Alternatively, [James Joyce](https://github.com/pjamesjoyce), who is the lead developer of [Futura](https://github.com/pjamesjoyce/futura), a software that handles and modifies LCA databases. That would be quite in line with the paper's topic.
