+++ setup.py
@@ -24,7 +24,7 @@ for dirpath, dirnames, filenames in os.walk("unfold"):
 
 def package_files(directory):
     paths = []
-    for (path, directories, filenames) in os.walk(directory):
+    for path, directories, filenames in os.walk(directory):
         for filename in filenames:
             paths.append(os.path.join("..", path, filename))
     return paths
