### 2023-02-22 #5050 @romainsacchi
@fraukewiese of course. I can further suggest:
Pr. Massimo Pizzol (AAU) @massimopizzol
Dr. Maxime Agez (CIRAIG) @MaximeAgez
Dr. Patrick Lamer (NREL) @plamers
