### 2023-02-23 #5050 @fraukewiese
@massimopizzol , @MaximeAgez , @IlkaCu  : – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review here is [unfold: removing the barriers to sharing and reproducing prospective life-cycle assessment databases](https://github.com/polca/unfold)
