### 2023-02-28 polca/unfold#1 @MaximeAgez
Problem folding an LCA database
Hi @romainsacchi,
I've been trying to fold two databases that I created in Activity-browser and an error popped up. I joined screenshots below:
![image](https://user-images.githubusercontent.com/24285636/221971167-9e9f1838-47b9-43e8-8d92-372eb53378a3.png)
![image](https://user-images.githubusercontent.com/24285636/221971324-f2390c26-5ad2-4482-82b1-a75927ac9239.png)
