### 2023-02-28 polca/unfold#1 @romainsacchi
I think the databases you are trying to fold contain datasets which miss some mandatory fields (in this case, the field unit). I will add a check routine tomorrow, to warn the user of missing fields and print a list of the guilty datasets. I will let you know.
