### 2023-02-28 polca/unfold#2 @MaximeAgez
API Documentation
90% of the functions/methods are documented, except in the unfold.py file where I would say the percentage drops to about 40-50% of functions/methods documented.
