### 2023-02-28 polca/unfold#3 @MaximeAgez
Lacking community guidelines
I think the package lacks the community guidelines required by JOSS: https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines
