### 2023-03-02 723e558
Add function to check for mandatory fields in database Add function to detect lack of common datasets between reference and scenario databases.
