+++ unfold/unfold.py
@@ -83,27 +83,47 @@ class Unfold:
         print("To unfold all scenarios, `unfold()`.")
         print("To unfold a specific scenario, `unfold(scenarios=[1,])`.")
 
-    def check_dependencies(self, dependencies: dict):
-        """Checks the dependencies."""
-        # we need to check that the source database exists
-        # and that the scenarios are compatible with the source database
+    def check_dependencies(self, dependencies: dict) -> None:
+        """
+        Checks the dependencies required to unfold the scenarios.
 
+        Parameters:
+        -----------
+        dependencies : dict
+            A dictionary containing the names of the databases required as dependencies.
+
+        Returns:
+        --------
+        None
+
+        Notes:
+        ------
+        This method checks that the required source database and scenarios are available and compatible with each other. If the dependencies are not provided, the method assumes that no dependencies are required.
+
+        If the required dependencies are not found in the available databases, the method prompts the user to input the database number for each dependency.
+
+        Once the user has provided the necessary information, the method updates the "source" attribute for each dependency with the name of the corresponding database.
+
+        """
+        # Get a list of available databases in the project
         available_databases = list(bw2data.databases.keys())
 
-        dependencies = dependencies or []
+        # If no dependencies are provided, set an empty dict
+        dependencies = dependencies or {}
 
-        if dependencies and all(
-            dependency in available_databases for dependency in dependencies.values()
-        ):
+        # Check if all dependencies are available in the project
+        if dependencies and all(dependency in available_databases for dependency in dependencies.values()):
+            # If all dependencies are available, update the "source" attribute for each dependency
             for database in self.dependencies:
                 database["source"] = dependencies[database["name"]]
         else:
-            # ask the user to give names to the databases
+            # If any dependencies are missing, prompt the user to input the necessary information
             print("The following databases are necessary to unfold the scenarios:")
 
             table = PrettyTable()
             table.field_names = ["No.", "Dependency", "System model", "Version"]
 
+            # Add a row to the table for each dependency
             for _db, database in enumerate(self.dependencies):
                 table.add_row(
                     [
@@ -113,6 +133,8 @@ class Unfold:
                         database.get("version", ""),
                     ]
                 )
+
+            # Print the table of required dependencies
             print(table)
             print("")
 
@@ -124,11 +146,15 @@ class Unfold:
                 "Database",
             ]
 
+            # Add a row to the table for each available database
             for _db, database in enumerate(available_databases):
                 table.add_row([_db + 1, database])
+
+            # Print the table of available databases
             print(table)
             print("")
 
+            # Prompt the user to input the database number for each dependency
             for _db, database in enumerate(self.dependencies):
                 db_number = input(
                     f"Indicate the database number for dependency {_db + 1}: "
@@ -136,8 +162,26 @@ class Unfold:
                 name = available_databases[int(db_number) - 1]
                 database["source"] = name
 
-    def build_mapping_for_dependencies(self, database):
-        """Builds a mapping for dependencies."""
+    def build_mapping_for_dependencies(self, database) -> None:
+        """
+        Builds a mapping for dependencies based on the given database.
+
+        Parameters:
+        -----------
+        database : list
+            A list of activity dictionaries representing a database.
+
+        Returns:
+        --------
+        None
+
+        Notes:
+        ------
+        This method builds a mapping for dependencies based on the given database. The mapping is stored in the "dependency_mapping" attribute of the current object.
+
[…]
