### 2023-03-02 c0669ea
Add sections to README.md to comply with JOSS' community guidelines.
