+++ unfold/__init__.py
@@ -1,5 +1,5 @@
 __all__ = ("Unfold", "Fold")
-__version__ = (1, 0, 2)
+__version__ = (1, 0, 3)
 
 from .fold import Fold
 from .unfold import Unfold
