+++ unfold/fold.py
@@ -24,7 +24,7 @@ from .data_cleaning import (
     check_mandatory_fields,
     get_biosphere_code,
     get_outdated_flows,
-    check_commonality_between_databases
+    check_commonality_between_databases,
 )
 
 DIR_DATAPACKAGE_TEMP = DATA_DIR / "temp"
@@ -134,8 +134,29 @@ class Fold:
         descriptions: List[str] = None,
     ):
         """
-        Identify the source database
-        :return: name of the source database
+        The identify_databases_to_fold function identifies the source database and the databases to be folded into the source database and extracts them.
+
+        :param source_database: Name of the source database to be used. If not specified, the user is prompted to choose from the available databases.
+        :param source_database_system_model: System model of the source database to be used. If not specified, the user is prompted to input.
+        :param source_database_version: Version of the source database to be used. If not specified, the user is prompted to input.
+        :param databases_to_fold: List of databases to be folded into the source database. If not specified, the user is prompted to input.
+        :param descriptions: Short descriptions of each database to be folded. If not specified, the user is prompted to input.
+        :return: The source dictionary containing information about the source database, such as its name, database, system model, and version.
+        :return: The databases_to_fold list containing dictionaries of information about each database to be folded, including its name, database, and description.
+
+        Functionality:
+
+        Checks whether the user has already specified a datapackage name and description. If not, prompts the user to input these details.
+        Lists the available databases and prompts the user to input the number of the reference database if the source_database input is not specified.
+        Identifies the dependencies of the source and folded databases.
+        Prompts the user to input the system model and version of the source database if not specified.
+        Prompts the user to input the list of databases to be folded and their descriptions if not specified.
+        Extracts the source database and ensures that mandatory fields are included.
+        Builds the mapping of dependencies for the source database.
+        Extracts each database to be folded, ensures that mandatory fields are included, and builds the mapping of dependencies.
+        Identifies whether any dependencies are external and, if so, extracts them.
+        Returns a set of the dependencies excluding the source and databases to be folded.
+
         """
 
         if not self.datapackage_name:
@@ -226,7 +247,6 @@ class Fold:
                 source_database_extracted, extracted_database, database
             )
 
-
             self.databases_to_fold.append(
                 {
                     "name": database,
@@ -290,9 +310,16 @@ class Fold:
 
     def get_list_unique_acts(self, scenarios: List[List[dict]]) -> list:
         """
-        Get a list of unique activities from a list of databases
-        :param scenarios: list of databases
-        :return: list of unique activities
+        Returns a list of unique activities from a list of databases, where each database is represented by a list of
+        datasets containing their respective exchanges.
+
+        :param scenarios: A list of databases, where each database is a list of datasets, with each dataset containing the
+                          exchanges.
+        :type scenarios: list
+
+        :return: A list of tuples representing the unique activities in the provided databases, where each tuple contains
+                 the activity name, reference product, location, categories, unit and type.
+        :rtype: list
         """
 
         list_unique_acts = []
@@ -343,6 +370,25 @@ class Fold:
         databases_to_fold: List[str] = None,
         descriptions: List[str] = None,
     ):
+        """
+        Folds one or more databases into a new package.
+
+        :param package_name: Name for the new datapackage.
+        :type package_name: str, optional
+        :param package_description: Short description for the new datapackage.
+        :type package_description: str, optional
+        :param source: Name of the source database.
+        :type source: str, optional
+        :param system_model: System model of the source database.
+        :type system_model: str, optional
+        :param version: Version of the source database.
+        :type version: float or str, optional
+        :param databases_to_fold: List of names of the databases to fold.
+        :type databases_to_fold: List[str], optional
[…]
