### 2023-03-02 polca/unfold#2 @romainsacchi
I have added documentation to the methods in `unfold.py`, as well as a bit in the other modules. See commit https://github.com/polca/unfold/commit/abb67a680506303173830c0f549cabb84744a0bb
