### 2023-03-02 polca/unfold#1 @romainsacchi
I think I may have solved the issue.
Can you try to update to `v.1.0.3` (from pip, or the latest from conda) and let me know?
