### 2023-03-02 polca/unfold#1 @romainsacchi
Hmm, it seems it was not pushed on conda, only pip.
