### 2023-03-02 polca/unfold#1 @romainsacchi
OK, it is on conda now as well.
