### 2023-03-02 polca/unfold#1 @MaximeAgez
I updated to 1.0.3. Now I face another Error:

![image](https://user-images.githubusercontent.com/24285636/222557074-e194ec29-eb6d-437d-bb37-04f4f25388dd.png)
![image](https://user-images.githubusercontent.com/24285636/222557098-6c672a17-d5c2-4e29-ab0b-35a9fcf45d11.png)

Seems like it does not think that my database is linked to ecoinvent. But it is linked to ecoinvent.

I tried with other databases in other projects with other versions of ecoinvent and the same error pops up.
