### 2023-03-03 polca/unfold#1 @romainsacchi
Hi @MaximeAgez ,

right, I think this is the expected behavior. The purpose of the library is to create a "mask" of a database that contains proprietary data. For example, let's say I made a copy of ecoinvent and modified some markets and some processes throughout the database. I cannot share my modified ecoinvent database, since 95% of it is still original ecoinvent data (as I modified only a few processes). `unfold` would create a "mask" (in the form of a data package) of that modified ecoinvent database that I can freely share (since it only contains differences between my modified database and the "reference" ecoinvent database). The users would apply this mask to their local original ecoinvent database, to reproduce my modified ecoinvent database.

So, in your case, I suspect "Glencore - Horne" is a database that simply links to the reference database (ecoinvent?), it is not a "modified" version of the reference database, since `unfold` could not detect any datasets in common.

The typical use I can think of would be when creating prospective LCA databases, which are often just modified ecoinvent databases where electricity markets have been changed, etc. This would allow reproducing these prospective LCA databases without sharing the databases themselves, but only the difference with respect to the unmodified database.

Hence, for example, we have [here](https://github.com/polca/replicate/blob/main/README.md) a list of data packages which one can unfold to reproduce locally some scenario-based LCA databases.

Do you think it deserves a better explanation in the paper/github repo?
