### 2023-03-03 polca/unfold#1 @MaximeAgez
Actually I reread the paper and it makes sense now. I misunderstood the purpose of the software but it's clearly stated, I was probably not as focused as I should have been.

I successfully folded and unfolded two premise-generated ecoinvent databases. But there is something that seems strange to me. Before folding the two databases had respectively 35549 and 34575 activities (it's ecoinvent_image_SSP2-Base_2023 and ecoinvent_image_SSP2-RCP19_2023, using ecoinvent3.8 cut-off).

![image](https://user-images.githubusercontent.com/24285636/222758094-8ce7a2a5-1ce6-4626-b67a-a7602cbe7045.png)

After unfolding them in another project, the 1st one has now 34575 activities:
![image](https://user-images.githubusercontent.com/24285636/222757954-35db3c34-dbac-41bc-91a5-ebf36c70ae63.png)

Is that an expected behaviour?
