### 2023-03-03 polca/unfold#1 @romainsacchi
It is indeed surprising.
Do you think you could help by running something like:

````

bw.project.set_current("premise")
list_act_db1 = [(d["name"], d["reference product"], d["location"]) for d in bw.Database("ecoinvent_image_SSP2-Base_2030")]

bw.project.set_current("ecoinvent3.8")
list_act_db2 = [(d["name"], d["reference product"], d["location"]) for d in bw.Database("ecoinvent_image_SSP2-Base_2030")]

print([d for d in list_act_db2 if d not in list_act_db1])

````
