### 2023-03-07 99577c1
Add `filter_out_datasets`which filters out datasets of a scenario that belong to other scenarios.
