### 2023-03-07 polca/unfold#1 @romainsacchi
Hi @MaximeAgez,
indeed, thanks for noticing that. These redundant datasets belong to other folded scenarios.
Although they are orphans and empty, having them there is not nice.
In a few minutes, v.1.0.4 should be released, preventing this.
