### 2023-03-12 polca/unfold#4 @mfastudillo
conflicting dependencies
unfold seems to be particularly useful when dealing with databases generated using premise but they seem to have conflicting requirements. unfold requires wurst==0.3.4 (https://github.com/polca/unfold/blob/main/requirements.txt) and premise wurst==0.3.3 (https://github.com/polca/premise/blob/master/requirements.txt)
