### 2023-03-12 polca/unfold#5 @mfastudillo
improve examples
I think it would be helpful to improve the example provided in the repository. The example seems to assume we already have a premise database in our computer. It think this could be improved adding instructions on how to generate such database with premise.

Otherwise a very simple example with fictitious databases would work too.
