### 2023-03-12 polca/unfold#5 @mfastudillo
A simple fictitious example could be used as pytest fixture to expand the test coverage
