### 2023-03-12 #5198 @mfastudillo
improve examples of use (https://github.com/polca/unfold/issues/5)
