### 2023-03-12 polca/unfold#6 @mfastudillo
statement of need
A question related to the JOSS paper. There seems to be alternatives to unfold. Instead of sharing the changes to the licensed database we share the code that modifies the licensed database or share recipes for changing it using something like Futura. To that respect, what is the advantage of using unfold ? I think that could help to clarify the contribution to the field.
