### 2023-03-14 67c7e52
Correct typos in paper.md. Add a paragraph to address issue #6.
