+++ paper.md
@@ -24,71 +24,78 @@ bibliography: paper.bib
 
 # Summary
 
-`unfold` is a Python package that allows reproducing life-cycle databases which
+`unfold` is a Python package that allows reproducing life-cycle databases which 
 partially build on a data source that cannot be shared. It produces 
-data packages that contain the differences between the databases to share and 
-the licensed data source. The data package also contains a metadata file that 
-describes the databases, the author, and other useful information. `unfold` allows to 
-pack and unpack any number of databases from a single data package, to ease the 
-sharing and reproducibility of prospective or scenario-based life-cycle assessment 
-databases.
+data packages that contain the differences between the databases to 
+share and the licensed data source. The data package also includes a 
+metadata file describing the databases, the author, and other helpful 
+information. `unfold` allows one to pack and unpack any databases from 
+a single data package to ease the sharing and reproducibility of prospective 
+or scenario-based life-cycle assessment databases.
+
 
 # Statement of need
 
-Life-cycle assessment (LCA) consists in the quantification of the environmental 
-footprint of a product or a process by considering impacts along each of its 
-relevant life-cycle phases [@ISO:2006]. It relies on databases describing 
-thousands of industrial processes, which interdependency and interactions 
-with the greater environment form large input-output matrices. 
-However, these matrices, often modified to fit the need of a study, 
+Life-cycle assessment (LCA) consists of quantifying the environmental footprint 
+of a product or a process by considering impacts along each of its relevant 
+life-cycle phases [@ISO:2006]. It relies on databases describing thousands of 
+industrial processes, which interdependency and interactions with the greater 
+environment form large input-output matrices. 
+However, these matrices -- often modified to fit the need of a study -- 
 are difficult to share when they contain data covered by a restrictive 
-end-user license agreement, such as that of the ecoinvent database [@Wernet:2016].
+End-user license agreement, such as that of the ecoinvent database [@Wernet:2016].
 
-The difficulty to share these databases prevents the reproducibility of results 
+The difficulty in sharing these databases prevents the reproducibility of results 
 and impedes the scientific validity of the work produced by the LCA community.
 
-Using the LCA framework `brightway2` [@Mutel:2017], `unfold` allows to share LCA 
-databases without exposing the data under license, sharing instead data packages used 
-to reproduce the databases locally (provided the end-users also have access to the 
-licensed data source).
+Using the LCA framework `brightway2` [@Mutel:2017], `unfold` allows sharing modified 
+LCA databases without:
 
-`unfold` is initially conceived to share LCA databases that have been heavily 
-modified, such as those generated with the premise package [@Sacchi:2022], or 
-those regularly produced within the field of prospective LCA (where the need to 
-modify extensively the source database is often required) – see the work of 
-Cox [@Cox:2018], Mendoza [@Mendoza:2018], Joyce [@Joyce:2022] and colleagues.
+* exposing the data under license,
+* going through the steps of data modification,
+
+but sharing instead data packages used to reproduce the databases locally (provided 
+the End-users also have access to the licensed data source).
+
+`unfold` is initially conceived to share heavily modified LCA databases, such as 
+those generated with the `premise` package [@Sacchi:2022] or those regularly produced 
+within the field of prospective LCA (where the need to modify the source database 
+extensively is often required) – see the work of Cox [@Cox:2018], Mendoza [@Mendoza:2018], 
+Joyce [@Joyce:2022] and colleagues.
+
 
 # Description
 
-LCA relies on a set of input-output matrices representing product and service exchanges
-that are typically sparse, with a density inferior to 1%. Sometimes, for the need of sensitivity
-or scenario analysis, up to hundreds of thousands of exchanges are added, modified, 
[…]
