### 2023-03-14 polca/unfold#6 @romainsacchi
Hi @mfastudillo
commit https://github.com/polca/unfold/commit/67c7e529bcc5072ff3f1cd986fdc5e4bf37b484a shows the modification brought to the manuscript.

Besides fixing some typos and improving the English, I added the following paragraph:

_Sharing scenario-based or prospective LCA databases is already possible through
other means, such as exchanging `Futura`'s recipes [@Joyce:2022] or
`premise`'s input parameters [@Sacchi:2022]. However, the diversity of shared
databases is eventually limited by the range of tools that generate them -- most
of the existing tools focus on modifying energy-intensive sectors.
On the other hand, `unfold` allows sharing databases regardless of the tool or
method used to generate them, as long as they present differences with the reference
database. For example, a database with manually adjusted leakage rates of ozone-depleting
gases can be shared just as easily._
