+++ conda/meta.yaml
@@ -22,7 +22,7 @@ requirements:
     - pandas
     - bw2io==0.8.7
     - bw2data==3.6.5
-    - wurst==0.3.4
+    - wurst>=0.3.3
     - prettytable
     - datapackage
     - sparse
