### 2023-03-14 polca/unfold#4 @romainsacchi
Hi @mfastudillo
commit https://github.com/polca/unfold/commit/9832fd361f2747ee47de7155feaea1c483e30e9c changes the version specification of `wurst` to `>=0.3.3`.
