+++ .github/workflows/main.yml
@@ -57,6 +57,7 @@ jobs:
         run: |
           pip install -r requirements.txt --upgrade pip
           pip install -e .
+          pip install brightway2
           pip install pytest
 
       - name: Run tests
