### 2023-03-14 polca/unfold#5 @mfastudillo
great, being picky here, but wouldn't it be better to call `bw2calc.LCA({bw2data.get_activity` instead of `bw.LCA({bw.get_activity` in the examples?. When I tried to run the example It resulted in errors because brightway2 was not installed (it is not part of the requirements). Everything worked fine when I used the imported bw2data and bw2calc
