### 2023-03-15 26ca65c
Use `bw2data` and `bw2calc` instead of `brightway2` in test and examples notebook.
