### 2023-03-16 #5198 @MaximeAgez
@fraukewiese I passed my comments to the authors in these issues (https://github.com/polca/unfold/issues/1#, https://github.com/polca/unfold/issues/2#, https://github.com/polca/unfold/issues/3#) and the changes made answered my comments.

I am satisfied with the submission. However, I do not have the necessary expertise on testing, so I left the cell empty.
