+++ CITATION.cff
@@ -1,0 +1,28 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Sacchi
+  given-names: Romain
+  orcid: "https://orcid.org/0000-0003-1440-0905"
+doi: 10.5281/zenodo.7777717
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Sacchi
+    given-names: Romain
+    orcid: "https://orcid.org/0000-0003-1440-0905"
+  date-published: 2023-03-29
+  doi: 10.21105/joss.05198
+  issn: 2475-9066
+  issue: 83
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 5198
+  title: "unfold: removing the barriers to sharing and reproducing
+    prospective life-cycle assessment databases"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.05198"
+  volume: 8
+title: "`unfold`: removing the barriers to sharing and reproducing
+  prospective life-cycle assessment databases"
